+++ CHANGELOG.rst
@@ -12,7 +12,7 @@ Changes
 * Empty string arguments to :code:`get_date_resolution` and :code:`parse_date` now raise :code:`ISOFormatError`, fixes `26 <https://bitbucket.org/nielsenb/aniso8601/issues/26/parse_date-parse_time-parse_datetime-fails>`_
 * None and non-string arguments to :code:`get_date_resolution` and :code:`parse_date` now raise :code:`ValueError`
 * Empty string arguments to :code:`parse_duration` now raise :code:`ISOFormatError`
-* None and non-string arguments to :code:`parse_duration` now raise :code:`ISOFormatError`
+* None and non-string arguments to :code:`parse_duration` now raise :code:`ValueError`
 * Empty string arguments to :code:`parse_interval` and :code:`parse_repeating_internval` now raise :code:`ISOFormatError`
 * None and non-string arguments to :code:`parse_interval` and :code:`parse_repeating_internval` now raise :code:`ValueError`
 * Empty string arguments to :code:`get_time_resolution` and :code:`parse_time` now raise :code:`ISOFormatError`